Rename the driver pkg to abb_hardware_interface. README updates.
