+++ abb_hardware_interface/abb_hardware_interface.xml
@@ -1,0 +1,9 @@
+<library path="abb_hardware_interface">
+  <class name="abb_hardware_interface/ABBSystemPositionOnlyHardware"
+         type="abb_hardware_interface::ABBSystemPositionOnlyHardware"
+         base_class_type="hardware_interface::SystemInterface">
+    <description>
+		ROS2 Control ABB driver.
+    </description>
+  </class>
+</library>
